refactor: reorganize lib folders for protocol acceleration libs
