+++ apps/ensindexer/src/api/handlers/resolution-api.ts
@@ -7,10 +7,10 @@ import { Hono } from "hono";
 
 import { errorResponse } from "@/api/lib/error-response";
 import { canAccelerateResolution } from "@/api/lib/indexing-status/can-accelerate-resolution";
-import { captureTrace } from "@/api/lib/protocol-tracing";
 import { resolveForward } from "@/api/lib/resolution/forward-resolution";
 import { resolvePrimaryNames } from "@/api/lib/resolution/multichain-primary-name-resolution";
 import { resolveReverse } from "@/api/lib/resolution/reverse-resolution";
+import { captureTrace } from "@/api/lib/tracing/protocol-tracing";
 import { validate } from "@/api/lib/validate";
 import { simpleMemoized } from "@/lib/simple-memoized";
 import { routes } from "@ensnode/ensnode-sdk/internal";
