+++ apps/ensindexer/src/api/index.ts
@@ -5,7 +5,7 @@ import schema from "ponder:schema";
 import { Hono } from "hono";
 import { cors } from "hono/cors";
 
-import { sdk } from "@/api/lib/instrumentation";
+import { sdk } from "@/api/lib/tracing/instrumentation";
 import config from "@/config";
 import { makeApiDocumentationMiddleware } from "@/lib/api-documentation";
 import { filterSchemaExtensions } from "@/lib/filter-schema-extensions";
