+++ apps/ensindexer/src/api/lib/protocol-acceleration/get-primary-name-from-index.ts
