+++ apps/ensindexer/src/api/lib/protocol-acceleration/get-records-from-index.ts
@@ -1,8 +1,8 @@
 import { db } from "ponder:api";
-import { onchainStaticResolverImplementsDefaultAddress } from "@/api/lib/acceleration/known-onchain-static-resolver";
-import type { IndexedResolverRecords } from "@/api/lib/make-records-response";
+import { onchainStaticResolverImplementsDefaultAddress } from "@/api/lib/protocol-acceleration/known-onchain-static-resolver";
+import type { IndexedResolverRecords } from "@/api/lib/resolution/make-records-response";
 import { withSpanAsync } from "@/lib/auto-span";
-import { makeResolverRecordsId } from "@/plugins/protocol-acceleration/lib/ids";
+import { makeResolverRecordsId } from "@/lib/protocol-acceleration/ids";
 import {
   ChainId,
   DEFAULT_EVM_COIN_TYPE,
