+++ apps/ensindexer/src/api/lib/resolution/reverse-resolution.ts
@@ -11,7 +11,7 @@ import {
 import { SpanStatusCode, trace } from "@opentelemetry/api";
 import { isAddress, isAddressEqual } from "viem";
 
-import { addProtocolStepEvent, withProtocolStepAsync } from "@/api/lib/protocol-tracing";
+import { addProtocolStepEvent, withProtocolStepAsync } from "@/api/lib/tracing/protocol-tracing";
 import { withActiveSpanAsync } from "@/lib/auto-span";
 import { resolveForward } from "./forward-resolution";
